back/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
Werkzeug==3.1.3
